Move NumericLiteral to its own module.
